chore: restructure goopstest code
