+++ goopstest/command_runner.go
@@ -8,34 +8,7 @@ import (
 	"time"
 
 	"github.com/gruyaume/goops"
-	"gopkg.in/yaml.v3"
 )
-
-type Context struct {
-	Charm         func() error
-	Metadata      goops.Metadata
-	AppName       string
-	UnitID        int
-	JujuVersion   string
-	ActionResults map[string]string
-	ActionError   error
-	JujuLog       []JujuLogLine
-	CharmErr      error
-}
-
-type LogLevel string
-
-const (
-	LogLevelInfo    LogLevel = "INFO"
-	LogLevelWarning LogLevel = "WARNING"
-	LogLevelError   LogLevel = "ERROR"
-	LogLevelDebug   LogLevel = "DEBUG"
-)
-
-type JujuLogLine struct {
-	Level   LogLevel
-	Message string
-}
 
 type fakeCommandRunner struct {
 	Command            string
@@ -944,238 +917,4 @@ func (f *fakeCommandRunner) handleActionGet(_ []string) {
 
 func (f *fakeCommandRunner) handleApplicationVersionSet(args []string) {
 	f.ApplicationVersion = args[0]
-}
-
-type fakeEnvGetter struct {
-	HookName    string
-	ActionName  string
-	Model       *Model
-	AppName     string
-	UnitID      int
-	JujuVersion string
-	Metadata    goops.Metadata
-}
-
-func (f *fakeEnvGetter) Get(key string) string {
-	switch key {
-	case "JUJU_HOOK_NAME":
-		return f.HookName
-	case "JUJU_ACTION_NAME":
-		return f.ActionName
-	case "JUJU_MODEL_NAME":
-		return f.Model.Name
-	case "JUJU_MODEL_UUID":
-		return f.Model.UUID
-	case "JUJU_UNIT_NAME":
-		return fmt.Sprintf("%s/%d", f.AppName, f.UnitID)
-	case "JUJU_VERSION":
-		return f.JujuVersion
-	}
-
-	return ""
-}
-
-func (f *fakeEnvGetter) ReadFile(name string) ([]byte, error) {
-	if strings.HasSuffix(name, "metadata.yaml") {
-		data, err := yaml.Marshal(f.Metadata)
-		if err != nil {
-			return nil, fmt.Errorf("failed to marshal metadata: %w", err)
-		}
-
-		return data, nil
-	}
-
-	return nil, fmt.Errorf("file %s not found", name)
-}
-
-// For each relation, we set the ID to: <name>:<number>
-func setRelationIDs(relations []*Relation) {
-	for i, relation := range relations {
-		if relation.ID == "" {
-			relation.ID = fmt.Sprintf("%s:%d", relation.Endpoint, i)
-		}
-	}
-}
-
-// For each relation, we set the remoteUnitsData so that it contains at leader 1 unit
-func setUnitIDs(relations []*Relation) {
-	for _, relation := range relations {
-		if relation.RemoteUnitsData == nil {
-			relation.RemoteUnitsData = make(map[UnitID]DataBag)
-		}
-
-		if len(relation.RemoteUnitsData) == 0 {
-			relation.RemoteUnitsData[UnitID(relation.RemoteAppName+"/0")] = DataBag{}
-		}
-	}
-}
-
-func (c *Context) Run(hookName string, state *State) (*State, error) {
-	if c.Charm == nil {
-		return nil, fmt.Errorf("charm function is not set in the context")
-	}
-
-	setRelationIDs(state.Relations)
-	setUnitIDs(state.Relations)
-
-	if state.Model == nil {
-		state.Model = &Model{
-			Name: "test-model",
-			UUID: "12345678-1234-5678-1234-567812345678",
-		}
-	}
-
-	fakeCommand := &fakeCommandRunner{
-		Output:      []byte(``),
-		Err:         nil,
-		Leader:      state.Leader,
-		Config:      state.Config,
-		Secrets:     state.Secrets,
-		Relations:   state.Relations,
-		Ports:       state.Ports,
-		StoredState: state.StoredState,
-		AppName:     c.AppName,
-		UnitID:      c.UnitID,
-	}
-
-	fakeEnv := &fakeEnvGetter{
-		HookName:    hookName,
-		Model:       state.Model,
-		AppName:     c.AppName,
-		UnitID:      c.UnitID,
-		JujuVersion: c.JujuVersion,
-		Metadata:    c.Metadata,
-	}
-
-	fakePebble := &fakePebbleGetter{
-		Containers: state.Containers,
-	}
-
-	goops.SetPebbleGetter(fakePebble)
-	goops.SetCommandRunner(fakeCommand)
-	goops.SetEnvGetter(fakeEnv)
-
-	err := c.Charm()
-	if err != nil {
-		c.CharmErr = err
-	}
-
-	state.UnitStatus = fakeCommand.UnitStatus
-	state.AppStatus = fakeCommand.AppStatus
-	state.Secrets = fakeCommand.Secrets
-	state.ApplicationVersion = fakeCommand.ApplicationVersion
-	state.Ports = fakeCommand.Ports
-	state.StoredState = fakeCommand.StoredState
-	state.Containers = fakePebble.Containers
-
-	c.JujuLog = fakeCommand.JujuLog
-
-	return state, nil
-}
-
-func (c *Context) RunAction(actionName string, state *State, params map[string]any) (*State, error) {
-	fakeCommandRunner := &fakeCommandRunner{
-		Output:           []byte(``),
-		Err:              nil,
-		Leader:           state.Leader,
-		Config:           state.Config,
-		Secrets:          state.Secrets,
-		ActionParameters: params,
-		StoredState:      state.StoredState,
-	}
-
-	if state.Model == nil {
-		state.Model = &Model{
-			Name: "test-model",
-			UUID: "12345678-1234-5678-1234-567812345678",
-		}
-	}
-
-	fakeEnvGetter := &fakeEnvGetter{
-		ActionName:  actionName,
-		Model:       state.Model,
-		AppName:     c.AppName,
-		UnitID:      c.UnitID,
-		JujuVersion: c.JujuVersion,
-	}
-
-	goops.SetCommandRunner(fakeCommandRunner)
-	goops.SetEnvGetter(fakeEnvGetter)
-
-	err := c.Charm()
-	if err != nil {
-		c.CharmErr = err
-	}
-
-	state.UnitStatus = fakeCommandRunner.UnitStatus
-	state.AppStatus = fakeCommandRunner.AppStatus
-	state.Secrets = fakeCommandRunner.Secrets
-	state.ApplicationVersion = fakeCommandRunner.ApplicationVersion
-	c.ActionResults = fakeCommandRunner.ActionResults
-	c.ActionError = fakeCommandRunner.ActionError
-
-	return state, nil
-}
-
-type Secret struct {
-	ID          string
-	Label       string
-	Content     map[string]string
-	Owner       string
-	Description string
-	Rotate      string
-	Expire      time.Time
-}
-
-type UnitID string
-
-type DataBag map[string]string
-
-type Relation struct {
-	Endpoint        string
-	Interface       string
-	ID              string
-	RemoteAppName   string
-	LocalAppData    DataBag
-	LocalUnitData   DataBag
-	RemoteAppData   DataBag
-	RemoteUnitsData map[UnitID]DataBag
-}
-
-type Port struct {
-	Port     int
-	Protocol string
-}
-
-type Model struct {
-	Name string
-	UUID string
-}
-
-type StoredState map[string]string
-
-type Mount struct {
-	Location string
-	Source   string
-}
-
-type Exec struct {
-	Command    []string
-	ReturnCode int
-	Stdout     string
-	Stderr     string
-}
-
-type State struct {
-	Leader             bool
-	UnitStatus         string
-	AppStatus          string
-	Config             map[string]any
-	Secrets            []*Secret
-	ApplicationVersion string
-	Relations          []*Relation
-	Ports              []*Port
-	Model              *Model
-	StoredState        StoredState
-	Containers         []*Container
 }
